credential filter library for scripts that send proxy requests on behalf of a user and that require transmission of security credentials (basic auth, cookies)
A script can query this library to know if a remote host can be trusted
to receive security credentials (basic auth, cookies), safe cookies
(as declared in the config file) or none of the above.

The script will also return the list of cookies into the equivalent
of multiple lines of HTTP Set-Cookie headers that can be fed into
a cookie jar used by the script.
According to a configuration file, a script